+++ javascript-modules/engines/astro-engine/lib/modules/content.js
@@ -1,0 +1,45 @@
+export const getCollection = (collectionKey, filter) => {
+  if (!window.__bookshop_collections) {
+    console.warn("[Bookshop] Failed to load site collections for live editing");
+    return [];
+  }
+
+  if (!window.__bookshop_collections[collectionKey]) {
+    console.warn("[Bookshop] Failed to load collection: ", collectionKey);
+    return [];
+  }
+
+  if (filter) {
+    return window.__bookshop_collections[collectionKey].filter(filter);
+  }
+  return window.__bookshop_collections[collectionKey];
+};
+
+export const getEntry = (...args) => {
+  if (args.length === 1) {
+    const { collection: collectionKey, slug: entrySlug, id: entryId } = args[0];
+    const collection = getCollection(collectionKey);
+    if (entryId) {
+      return collection.find(({ id }) => id === entryId);
+    } else if (entrySlug) {
+      return collection.find(({ slug }) => slug === entrySlug);
+    }
+    return console.warn(
+      "[Bookshop] Failed to load entries, invalid arguments: ",
+      args
+    );
+  }
+
+  const [collectionKey, entryKey] = args;
+  const collection = getCollection(collectionKey);
+
+  return collection.find(({ id, slug }) => entryKey === (slug ?? id));
+};
+
+export const getEntries = (entries) => {
+  return entries.map(getEntry);
+};
+
+export const getEntryBySlug = (collection, slug) => {
+  return getEntry({ collection, slug });
+};